Outsource PrintCoriolisCoefficients() and fix an annoying bug
    bugfix:
    The Prototype of GetSettingsGetopt() was of type struct settings
    instead of void, resulting in very weird behaviour including a
    sporadic segmentation fault...

    Move the output of coriolis coefficients to its own function

+++ TheWatson/input/PreprocessBuffer.c
@@ -11,7 +11,7 @@ char * PreprocessBuffer(char* inputfile, int linenumber, char* buffer, const cha
 
 // check if line is complete (search for newline char)
     int control = 0;
-    for(long unsigned i = 0; i < strlen(buffer); ++i){
+    for(size_t i = 0; i < strlen(buffer); ++i){
         if(buffer[i] == '\n') { control = 1; }
     }
     if(control == 0){
